Tests: Move some support functions to separate utilities library
